+++ eras/shelley/impl/CHANGELOG.md
@@ -51,6 +51,21 @@
 
 ### `testlib`
 
+* Add:
+  - `submitBlock_`
+  - `submitBlock`
+  - `submitFailingBlock`
+  - `submitFailingBlockM`
+  - `withTxsInBlock_`
+  - `withTxsInBlock`
+  - `withTxsInFailingBlock`
+  - `withTxsInFailingBlockM`
+  - `tryTxsInBlock`
+* Remove `tryRunImpBBODY`
+* Add `Eq` instances for:
+  - `AlonzoBbodyEvent`
+  - `ShelleyBbodyEvent`
+  - `ShelleyLedgersEvent`
 * Add `NFData` and `ToExpr` constraints and instances for:
   - `AlonzoBlockBody`
   - `AlonzoBbodyPredFailure`
@@ -60,7 +75,7 @@
   - `BHeaderView`
   - `Block`
 * Add a `Generic` instance for `BHeaderView`
-* Add `impEventsFrom`
+* Add `impEventsFrom`, `impTransactionsFrom`
 * Change type of `ImpTestState.impEvents` field from `[]` to `Seq`
 * Remove `huddle-cddl` and the `CDDL` modules.
 * Add `ToCBOR (StashedAVVMAddresses era)` superclass to `ShelleyEraTest`